Checkpoint.  Controller can share validators with Job and render results to user
