--- db/migrate/20210212173631_create_vba_documents_upload_files.rb
@@ -1,9 +1,0 @@
-class CreateVBADocumentsUploadFiles < ActiveRecord::Migration[6.0]
-  def change
-    create_table(:vba_documents_upload_files) do |t|
-      t.string :guid
-      t.timestamps null: false
-    end
-    add_index(:vba_documents_upload_files, [:guid])
-  end
-end
