+++ db/migrate/20210215155459_add_column_active_storage_to_upload_submission.rb
@@ -1,0 +1,5 @@
+class AddColumnActiveStorageToUploadSubmission < ActiveRecord::Migration[6.0]
+  def change
+    add_column :vba_documents_upload_submissions, :use_active_storage, :boolean, default: false
+  end
+end
